add shadow chain  governance and ongx contract
