+++ core/genesis/genesis.go
@@ -34,7 +34,7 @@ import (
 	"github.com/ontio/ontology/core/utils"
 	"github.com/ontio/ontology/smartcontract/service/native/global_params"
 	"github.com/ontio/ontology/smartcontract/service/native/governance"
-	"github.com/ontio/ontology/smartcontract/service/native/ont"
+	"github.com/ontio/ontology/smartcontract/service/native/ongx"
 	nutils "github.com/ontio/ontology/smartcontract/service/native/utils"
 	"github.com/ontio/ontology/smartcontract/service/neovm"
 )
@@ -45,10 +45,8 @@ const (
 )
 
 var (
-	ONTToken   = newGoverningToken()
-	ONGToken   = newUtilityToken()
-	ONTTokenID = ONTToken.Hash()
-	ONGTokenID = ONGToken.Hash()
+	ONGXToken   = newUtilityToken()
+	ONGXTokenID = ONGXToken.Hash()
 )
 
 var GenBlockTime = (config.DEFAULT_GEN_BLOCK_TIME * time.Second)
@@ -92,8 +90,7 @@ func BuildGenesisBlock(defaultBookkeeper []keypair.PublicKey, genesisConfig *con
 	}
 
 	//block
-	ont := newGoverningToken()
-	ong := newUtilityToken()
+	ongx := newUtilityToken()
 	param := newParamContract()
 	oid := deployOntIDContract()
 	auth := deployAuthContract()
@@ -102,13 +99,11 @@ func BuildGenesisBlock(defaultBookkeeper []keypair.PublicKey, genesisConfig *con
 	genesisBlock := &types.Block{
 		Header: genesisHeader,
 		Transactions: []*types.Transaction{
-			ont,
-			ong,
+			ongx,
 			param,
 			oid,
 			auth,
 			config,
-			newGoverningInit(),
 			newUtilityInit(),
 			newParamInit(),
 			govConfig,
@@ -116,16 +111,6 @@ func BuildGenesisBlock(defaultBookkeeper []keypair.PublicKey, genesisConfig *con
 	}
 	genesisBlock.RebuildMerkleRoot()
 	return genesisBlock, nil
-}
-
-func newGoverningToken() *types.Transaction {
-	mutable := utils.NewDeployTransaction(nutils.OntContractAddress[:], "ONT", "1.0",
-		"Ontology Team", "contact@ont.io", "Ontology Network ONT Token", true)
-	tx, err := mutable.IntoImmutable()
-	if err != nil {
-		panic("constract genesis governing token transaction error ")
-	}
-	return tx
 }
 
 func newUtilityToken() *types.Transaction {
@@ -179,43 +164,11 @@ func deployOntIDContract() *types.Transaction {
 	return tx
 }
 
-func newGoverningInit() *types.Transaction {
-	bookkeepers, _ := config.DefConfig.GetBookkeepers()
-
-	var addr common.Address
-	if len(bookkeepers) == 1 {
-		addr = types.AddressFromPubKey(bookkeepers[0])
-	} else {
-		m := (5*len(bookkeepers) + 6) / 7
-		temp, err := types.AddressFromMultiPubKeys(bookkeepers, m)
-		if err != nil {
-			panic(fmt.Sprint("wrong bookkeeper config, caused by", err))
-		}
-		addr = temp
-	}
-
-	distribute := []struct {
-		addr  common.Address
-		value uint64
-	}{{addr, constants.ONT_TOTAL_SUPPLY}}
-
-	args := bytes.NewBuffer(nil)
-	nutils.WriteVarUint(args, uint64(len(distribute)))
-	for _, part := range distribute {
-		nutils.WriteAddress(args, part.addr)
-		nutils.WriteVarUint(args, part.value)
-	}
-
-	mutable := utils.BuildNativeTransaction(nutils.OntContractAddress, ont.INIT_NAME, args.Bytes())
-	tx, err := mutable.IntoImmutable()
-	if err != nil {
-		panic("constract genesis governing token transaction error ")
-	}
-	return tx
-}
-
 func newUtilityInit() *types.Transaction {
-	mutable := utils.BuildNativeTransaction(nutils.OngContractAddress, ont.INIT_NAME, []byte{})
+	bf := new(bytes.Buffer)
+	nutils.WriteAddress(bf, genAdminAddress())
+	mutable := utils.BuildNativeTransaction(nutils.OngContractAddress, ongx.INIT_NAME, bf.Bytes())
+
 	tx, err := mutable.IntoImmutable()
 	if err != nil {
 		panic("constract genesis governing token transaction error ")
@@ -240,9 +193,20 @@ func newParamInit() *types.Transaction {
 	for _, v := range s {
 		params.SetParam(global_params.Param{Key: v, Value: INIT_PARAM[v]})
 	}
+
 	bf := new(bytes.Buffer)
 	params.Serialize(bf)
+	nutils.WriteAddress(bf, genAdminAddress())
 
+	mutable := utils.BuildNativeTransaction(nutils.ParamContractAddress, global_params.INIT_NAME, bf.Bytes())
+	tx, err := mutable.IntoImmutable()
+	if err != nil {
+		panic("constract genesis governing token transaction error ")
+	}
+	return tx
+}
+
+func genAdminAddress() common.Address {
 	bookkeepers, _ := config.DefConfig.GetBookkeepers()
 	var addr common.Address
 	if len(bookkeepers) == 1 {
@@ -255,14 +219,7 @@ func newParamInit() *types.Transaction {
 		}
 		addr = temp
 	}
-	nutils.WriteAddress(bf, addr)
-
-	mutable := utils.BuildNativeTransaction(nutils.ParamContractAddress, global_params.INIT_NAME, bf.Bytes())
-	tx, err := mutable.IntoImmutable()
-	if err != nil {
-		panic("constract genesis governing token transaction error ")
-	}
-	return tx
+	return addr
 }
 
 func newGoverConfigInit(config []byte) *types.Transaction {
