+++ smartcontract/service/native/ongx/states.go
@@ -16,7 +16,7 @@
  * along with The ontology.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ont
+package ongx
 
 import (
 	"fmt"
@@ -213,3 +213,54 @@ func (this *TransferFrom) Deserialization(source *common.ZeroCopySource) error {
 
 	return err
 }
+
+type Inflations struct {
+	Inflations []Swap
+}
+
+func (this *Inflations) Serialize(sink *common.ZeroCopySink) {
+	utils.EncodeVarUint(sink, uint64(len(this.Inflations)))
+	for _, v := range this.Inflations {
+		v.Serialize(sink)
+	}
+}
+
+func (this *Inflations) Deserialize(source *common.ZeroCopySource) error {
+	n, err := utils.DecodeVarUint(source)
+	if err != nil {
+		return fmt.Errorf("Inflations deserialize count error:%s", err)
+	}
+	for i := 0; uint64(i) < n; i++ {
+		var swap Swap
+		if err := swap.Deserialize(source); err != nil {
+			return err
+		}
+		this.Inflations = append(this.Inflations, swap)
+	}
+	return nil
+}
+
+type Swap struct {
+	Addr    common.Address
+	Value uint64
+}
+
+func (this *Swap) Serialize(sink *common.ZeroCopySink) {
+	utils.EncodeAddress(sink, this.Addr)
+	utils.EncodeVarUint(sink, this.Value)
+}
+
+func (this *Swap) Deserialize(source *common.ZeroCopySource) error {
+	var err error
+	this.Addr, err = utils.DecodeAddress(source)
+	if err != nil {
+		return fmt.Errorf("Inflation deserialize to error:%s", err)
+	}
+	this.Value, err = utils.DecodeVarUint(source)
+	if err != nil {
+		fmt.Errorf("Inflation deserialize value error:%s", err)
+	}
+	return nil
+}
+
+
